MGMT-6812: Move isoeditor test global state into a suite file
Previously the test for creating the iso template was actually editing
the shared iso file created in the BeforeSuite function. Removing this
caused the test for creating the cluster iso to fail.

This commit makes this shared state clear by moving the suite functions
to a test suite file as well as removing the test "contamination" by
moving the offset embed call into the BeforeSuite function.
