+++ internal/isoeditor/isoeditor_suite_test.go
@@ -1,0 +1,86 @@
+package isoeditor
+
+import (
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestIsoEditor(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "IsoEditor")
+}
+
+const (
+	volumeID      = "Assisted123"
+	testRootFSURL = "https://example.com/pub/openshift-v4/dependencies/rhcos/4.7/4.7.7/rhcos-live-rootfs.x86_64.img"
+)
+
+var (
+	isoDir   string
+	isoFile  string
+	filesDir string
+)
+
+var _ = BeforeSuite(func() {
+	filesDir, isoDir, isoFile = createIsoViaGenisoimage(volumeID)
+	err := embedOffsetsInSystemArea(isoFile)
+	Expect(err).ToNot(HaveOccurred())
+})
+
+var _ = AfterSuite(func() {
+	os.RemoveAll(filesDir)
+	os.RemoveAll(isoDir)
+})
+
+func createIsoViaGenisoimage(volumeID string) (string, string, string) {
+	grubConfig := `
+menuentry 'RHEL CoreOS (Live)' --class fedora --class gnu-linux --class gnu --class os {
+	linux /images/pxeboot/vmlinuz random.trust_cpu=on rd.luks.options=discard coreos.liveiso=rhcos-46.82.202010091720-0 ignition.firstboot ignition.platform.id=metal
+	initrd /images/pxeboot/initrd.img /images/ignition.img
+}
+`
+	isoLinuxConfig := `
+label linux
+  menu label ^RHEL CoreOS (Live)
+  menu default
+  kernel /images/pxeboot/vmlinuz
+  append initrd=/images/pxeboot/initrd.img,/images/ignition.img random.trust_cpu=on rd.luks.options=discard coreos.liveiso=rhcos-46.82.202010091720-0 ignition.firstboot ignition.platform.id=metal
+`
+
+	filesDir, err := ioutil.TempDir("", "isotest")
+	Expect(err).ToNot(HaveOccurred())
+
+	isoDir, err := ioutil.TempDir("", "isotest")
+	Expect(err).ToNot(HaveOccurred())
+	isoPath := filepath.Join(isoDir, "test.iso")
+
+	err = os.Mkdir(filepath.Join(filesDir, "files"), 0755)
+	Expect(err).ToNot(HaveOccurred())
+	err = os.MkdirAll(filepath.Join(filesDir, "files/images/pxeboot"), 0755)
+	Expect(err).ToNot(HaveOccurred())
+	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/pxeboot/rootfs.img"), []byte("this is rootfs"), 0600)
+	Expect(err).ToNot(HaveOccurred())
+	err = os.MkdirAll(filepath.Join(filesDir, "files/EFI/redhat"), 0755)
+	Expect(err).ToNot(HaveOccurred())
+	err = ioutil.WriteFile(filepath.Join(filesDir, "files/EFI/redhat/grub.cfg"), []byte(grubConfig), 0600)
+	Expect(err).ToNot(HaveOccurred())
+	err = os.MkdirAll(filepath.Join(filesDir, "files/isolinux"), 0755)
+	Expect(err).ToNot(HaveOccurred())
+	err = ioutil.WriteFile(filepath.Join(filesDir, "files/isolinux/isolinux.cfg"), []byte(isoLinuxConfig), 0600)
+	Expect(err).ToNot(HaveOccurred())
+	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/assisted_installer_custom.img"), make([]byte, RamDiskPaddingLength), 0600)
+	Expect(err).ToNot(HaveOccurred())
+	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/ignition.img"), make([]byte, IgnitionPaddingLength), 0600)
+	Expect(err).ToNot(HaveOccurred())
+	cmd := exec.Command("genisoimage", "-rational-rock", "-J", "-joliet-long", "-V", volumeID, "-o", isoPath, filepath.Join(filesDir, "files"))
+	err = cmd.Run()
+	Expect(err).ToNot(HaveOccurred())
+
+	return filesDir, isoDir, isoPath
+}
