+++ internal/isoeditor/rhcos_test.go
@@ -6,10 +6,7 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
-	"os/exec"
-	"path/filepath"
 	"strings"
-	"testing"
 
 	"github.com/cavaliercoder/go-cpio"
 	"github.com/golang/mock/gomock"
@@ -21,40 +18,18 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-const (
-	testRootFSURL = "https://example.com/pub/openshift-v4/dependencies/rhcos/4.7/4.7.7/rhcos-live-rootfs.x86_64.img"
-)
-
 func getTestLog() logrus.FieldLogger {
 	l := logrus.New()
 	l.SetOutput(ioutil.Discard)
 	return l
 }
 
-func TestIsoEditor(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "IsoEditor")
-}
-
 var _ = Context("with test files", func() {
 	var (
-		isoDir                  string
-		isoFile                 string
-		filesDir                string
 		workDir                 string
-		volumeID                = "Assisted123"
 		ctrl                    *gomock.Controller
 		mockStaticNetworkConfig *staticnetworkconfig.MockStaticNetworkConfig
 	)
-
-	BeforeSuite(func() {
-		filesDir, isoDir, isoFile = createIsoViaGenisoimage(volumeID)
-	})
-
-	AfterSuite(func() {
-		os.RemoveAll(filesDir)
-		os.RemoveAll(isoDir)
-	})
 
 	BeforeEach(func() {
 		var err error
@@ -72,8 +47,6 @@ var _ = Context("with test files", func() {
 	Describe("CreateMinimalISOTemplate", func() {
 		It("iso created successfully", func() {
 			editor := editorForFile(isoFile, workDir, mockStaticNetworkConfig)
-			err := embedOffsetsInSystemArea(isoFile)
-			Expect(err).ToNot(HaveOccurred())
 			file, err := editor.CreateMinimalISOTemplate(testRootFSURL)
 			Expect(err).ToNot(HaveOccurred())
 
@@ -267,53 +240,6 @@ func editorForFile(iso string, workDir string, staticNetworkConfig staticnetwork
 		isoHandler: isoutil.NewHandler(iso, workDir),
 		log:        getTestLog(),
 	}
-}
-
-func createIsoViaGenisoimage(volumeID string) (string, string, string) {
-	grubConfig := `
-menuentry 'RHEL CoreOS (Live)' --class fedora --class gnu-linux --class gnu --class os {
-	linux /images/pxeboot/vmlinuz random.trust_cpu=on rd.luks.options=discard coreos.liveiso=rhcos-46.82.202010091720-0 ignition.firstboot ignition.platform.id=metal
-	initrd /images/pxeboot/initrd.img /images/ignition.img
-}
-`
-	isoLinuxConfig := `
-label linux
-  menu label ^RHEL CoreOS (Live)
-  menu default
-  kernel /images/pxeboot/vmlinuz
-  append initrd=/images/pxeboot/initrd.img,/images/ignition.img random.trust_cpu=on rd.luks.options=discard coreos.liveiso=rhcos-46.82.202010091720-0 ignition.firstboot ignition.platform.id=metal
-`
-
-	filesDir, err := ioutil.TempDir("", "isotest")
-	Expect(err).ToNot(HaveOccurred())
-
-	isoDir, err := ioutil.TempDir("", "isotest")
-	Expect(err).ToNot(HaveOccurred())
-	isoFile := filepath.Join(isoDir, "test.iso")
-
-	err = os.Mkdir(filepath.Join(filesDir, "files"), 0755)
-	Expect(err).ToNot(HaveOccurred())
-	err = os.MkdirAll(filepath.Join(filesDir, "files/images/pxeboot"), 0755)
-	Expect(err).ToNot(HaveOccurred())
-	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/pxeboot/rootfs.img"), []byte("this is rootfs"), 0600)
-	Expect(err).ToNot(HaveOccurred())
-	err = os.MkdirAll(filepath.Join(filesDir, "files/EFI/redhat"), 0755)
-	Expect(err).ToNot(HaveOccurred())
-	err = ioutil.WriteFile(filepath.Join(filesDir, "files/EFI/redhat/grub.cfg"), []byte(grubConfig), 0600)
-	Expect(err).ToNot(HaveOccurred())
-	err = os.MkdirAll(filepath.Join(filesDir, "files/isolinux"), 0755)
-	Expect(err).ToNot(HaveOccurred())
-	err = ioutil.WriteFile(filepath.Join(filesDir, "files/isolinux/isolinux.cfg"), []byte(isoLinuxConfig), 0600)
-	Expect(err).ToNot(HaveOccurred())
-	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/assisted_installer_custom.img"), make([]byte, RamDiskPaddingLength), 0600)
-	Expect(err).ToNot(HaveOccurred())
-	err = ioutil.WriteFile(filepath.Join(filesDir, "files/images/ignition.img"), make([]byte, IgnitionPaddingLength), 0600)
-	Expect(err).ToNot(HaveOccurred())
-	cmd := exec.Command("genisoimage", "-rational-rock", "-J", "-joliet-long", "-V", volumeID, "-o", isoFile, filepath.Join(filesDir, "files"))
-	err = cmd.Run()
-	Expect(err).ToNot(HaveOccurred())
-
-	return filesDir, isoDir, isoFile
 }
 
 func validateFileContainsLine(filename string, content string) {
